requirements.txt
pyttsx3==2.90
SpeechRecognition==3.10.3
llama-cpp-python